done deleteTest functionality if delete test. test going to sampletests history and modify fucntionly of delete sample. if delete sample all tests from sample are deleted and going into sampletestshistory

+++ server/samplesMicroService/services/sampleTestsServices.js
@@ -61,6 +61,16 @@ const updateTestFromSample = async(sampleId, data, testId) => {
 
 
 const deleteTestFromSample = async (sampleId, testId) => {
+
+    const existingTest = await sampleTestsModel.getTestByIdFromSample(sampleId,testId)
+
+    if (!existingTest) {
+        throw new Error('this test doesnt exist')
+    }
+
+    const deleteAction = 'deleteTest'
+    await sampleTestsModel.insertIntoSampleHistoryUpdateAndDelete(testId,deleteAction,existingTest,null,existingTest.performedby,sampleId)
+
     return await sampleTestsModel.deleteTestFromSample(sampleId,testId)
    
 }
